# docs/swagger.yaml
basePath: /api
definitions:
...
  internal_handlers.ErrorResponse:
    properties:
      error:
        type: string
    required:
    - error
    type: object
...
  internal_handlers.ProcedureCategoryCount:
    properties:
      count:
        type: integer
      name:
        type: string
      tp:
        type: integer
    type: object
...
  internal_handlers.PublicPetInfo:
    properties:
      birth:
        type: string
      breed:
        type: string
      castrated:
        type: boolean
      chip:
        type: string
      color:
        type: string
      id:
        type: string
      name:
        type: string
      sex:
        type: string
      species:
        type: string
    type: object
  internal_handlers.PublicPetResponse:
    properties:
      categories:
        items:
          $ref: '#/definitions/internal_handlers.ProcedureCategoryCount'
        type: array
      pet:
        $ref: '#/definitions/internal_handlers.PublicPetInfo'
    type: object
  internal_handlers.PublicProcedureItem:
    properties:
      date:
        type: string
      diagnosis:
        type: string
      id:
        type: string
      name:
        type: string
      nextDate:
        type: string
      notes:
        type: string
      vetName:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: Veterinary clinic management API
  title: VetApp API
  version: "1.0"
paths:
...
  /public/pets/{id}:
    get:
      parameters:
      - description: Pet ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handlers.PublicPetResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
      summary: Get public pet profile
      tags:
      - public
  /public/pets/{id}/procedures:
    get:
      parameters:
      - description: Pet ID
        in: path
        name: id
        required: true
        type: integer
      - description: Procedure type
        in: query
        name: tp
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/internal_handlers.PublicProcedureItem'
            type: array
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
      summary: Get public pet procedures
      tags:
      - public
  /public/pets/code/{code}:
    get:
      parameters:
      - description: 4-digit access code
        in: path
        name: code
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handlers.PublicPetResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
      summary: Look up pet by PIN code
      tags:
      - public
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"