# internal/handlers/public.go
package handlers

import (
	"net/http"
	"strconv"

	"vetapp-backend/internal/models"

	"github.com/go-chi/chi/v5"
	"gorm.io/gorm"
)

// PublicHandler handles unauthenticated public endpoints.
type PublicHandler struct {
	db *gorm.DB
}

// NewPublicHandler creates a new PublicHandler.
func NewPublicHandler(db *gorm.DB) *PublicHandler {
	return &PublicHandler{db: db}
}

// --- Response types ---

// PublicPetInfo is the public view of a pet (no sensitive data).
type PublicPetInfo struct {
	ID        string  `json:"id"`
	Name      string  `json:"name"`
	Species   string  `json:"species"`
	Breed     string  `json:"breed"`
	Sex       string  `json:"sex"`
	Chip      string  `json:"chip"`
	Birth     *string `json:"birth"`
	Color     string  `json:"color"`
	Castrated bool    `json:"castrated"`
}

// ProcedureCategoryCount is a procedure type with its record count.
type ProcedureCategoryCount struct {
	TP    int    `json:"tp"`
	Name  string `json:"name"`
	Count int    `json:"count"`
}

// PublicPetResponse is the full public pet profile.
type PublicPetResponse struct {
	Pet        PublicPetInfo           `json:"pet"`
	Categories []ProcedureCategoryCount `json:"categories"`
}

// procedureTypeNames maps TP codes to canonical Georgian display names.
//
// This mapping must mirror the legacy PHP app's scheme — see
// `ProcedureHandler.Types()` for the source-of-truth comment block. Any
// new tp added here must also be added in `procedureNameForTP` (used for
// auto-populating `tpname` on owner-created records) and in the mobile
// app's `CATEGORY_TP` constant.
var procedureTypeNames = map[int]string{
	1:   "ვაქცინაცია",
	2:   "ანალიზი (ძაღლი)",
	22:  "ანალიზი (კატა)",
	222: "ანალიზი (სხვა)",
	11:  "ექტოპარაზიტების პრევენცია",
	12:  "დეჰელმინთიზაცია",
	101: "სტომატოლოგია",
	102: "კარდიოლოგია",
	103: "ოქსიგენოთერაპია",
	104: "ტრავმატოლოგია",
	105: "დერმატოლოგია",
	106: "ქირურგია",
	107: "სხვა პროცედურა",
	108: "კონსულტაცია",
	109: "რადიოლოგია",
	110: "სტერილიზაცია/კასტრაცია",
	115: "მიკროჩიპი",
	116: "ლაბორატორია",
	202: "თერაპია",
	203: "ოფთალმოლოგია",
}

// GetPet returns public pet info and procedure category counts.
// @Summary Get public pet profile
// @Tags public
// @Produce json
// @Param id path int true "Pet ID"
// @Success 200 {object} PublicPetResponse
// @Failure 404 {object} ErrorResponse
// @Router /public/pets/{id} [get]
func (h *PublicHandler) GetPet(w http.ResponseWriter, r *http.Request) {
	id := chi.URLParam(r, "id")

	var pet models.Pet
	if err := h.db.First(&pet, id).Error; err != nil {
		writeJSON(w, http.StatusNotFound, ErrorResponse{Error: "pet not found"})
		return
	}

	h.writePetProfile(w, pet)
}

// LookupByCode returns public pet info by the 4-digit access code.
// @Summary Look up pet by PIN code
// @Tags public
// @Produce json
// @Param code path string true "4-digit access code"
// @Success 200 {object} PublicPetResponse
// @Failure 404 {object} ErrorResponse
// @Router /public/pets/code/{code} [get]
func (h *PublicHandler) LookupByCode(w http.ResponseWriter, r *http.Request) {
	code := chi.URLParam(r, "code")
	if len(code) != 4 {
		writeJSON(w, http.StatusNotFound, ErrorResponse{Error: "pet not found"})
		return
	}

	var pet models.Pet
	if err := h.db.Where("code = ?", code).First(&pet).Error; err != nil {
		writeJSON(w, http.StatusNotFound, ErrorResponse{Error: "pet not found"})
		return
	}

	h.writePetProfile(w, pet)
}

// PublicProcedureItem is a single procedure in the public view.
type PublicProcedureItem struct {
	ID       string  `json:"id"`
	Date     string  `json:"date"`
	NextDate string  `json:"nextDate,omitempty"`
	Name     string  `json:"name,omitempty"`
	Diagnosis string `json:"diagnosis,omitempty"`
	Notes    string  `json:"notes,omitempty"`
	VetName  string  `json:"vetName,omitempty"`
}

// GetPetProcedures returns public procedure list for a pet, filtered by tp.
// @Summary Get public pet procedures
// @Tags public
// @Produce json
// @Param id path int true "Pet ID"
// @Param tp query int false "Procedure type"
// @Success 200 {array} PublicProcedureItem
// @Failure 404 {object} ErrorResponse
// @Router /public/pets/{id}/procedures [get]
func (h *PublicHandler) GetPetProcedures(w http.ResponseWriter, r *http.Request) {
	id := chi.URLParam(r, "id")

	var pet models.Pet
	if err := h.db.First(&pet, id).Error; err != nil {
		writeJSON(w, http.StatusNotFound, ErrorResponse{Error: "pet not found"})
		return
	}

	petID := strconv.Itoa(int(pet.ID))
	tpStr := r.URL.Query().Get("tp")

	// Special case: tp=999 returns allergies
	if tpStr == "999" {
		var allergies []models.Allergy
		h.db.Where("uuid = ?", petID).Order("id DESC").Find(&allergies)

		items := make([]PublicProcedureItem, 0, len(allergies))
		for _, a := range allergies {
			items = append(items, PublicProcedureItem{
				ID:        strconv.Itoa(int(a.ID)),
				Date:      a.Date,
				Name:      a.Name,
				Notes:     "",
			})
		}
		writeJSON(w, http.StatusOK, items)
		return
	}

	query := h.db.Where("uuid = ?", petID)
	if tpStr != "" {
		query = query.Where("tp = ?", tpStr)
	}

	var procs []models.Procedure
	query.Order("date DESC, id DESC").Find(&procs)

	items := make([]PublicProcedureItem, 0, len(procs))
	for _, p := range procs {
		item := PublicProcedureItem{
			ID:        strconv.Itoa(int(p.ID)),
			Date:      p.Date,
			NextDate:  p.Date2,
			Name:      p.Vac,
			Diagnosis: p.Diagn,
			Notes:     p.Nout,
			VetName:   p.PName,
		}
		items = append(items, item)
	}

	writeJSON(w, http.StatusOK, items)
}

// writePetProfile builds and writes the public pet profile response.
func (h *PublicHandler) writePetProfile(w http.ResponseWriter, pet models.Pet) {
	petID := strconv.Itoa(int(pet.ID))

	info := PublicPetInfo{
		ID:        petID,
		Name:      pet.Name,
		Species:   pet.Pet,
		Breed:     pet.Variety,
		Sex:       pet.Sex,
		Chip:      pet.Chip,
		Color:     pet.Color,
		Castrated: pet.Cast != "",
	}
	if pet.Date != "" {
		info.Birth = &pet.Date
	}

	// Count procedures by type
	type tpCount struct {
		TP    int
		Count int
	}
	var counts []tpCount
	h.db.Model(&models.Procedure{}).
		Select("tp as tp, COUNT(*) as count").
		Where("uuid = ?", petID).
		Group("tp").
		Scan(&counts)

	// Count allergies
	var allergyCount int64
	h.db.Model(&models.Allergy{}).Where("uuid = ?", petID).Count(&allergyCount)

	// Build categories — include all types that have records
	categories := make([]ProcedureCategoryCount, 0, len(counts)+1)
	for _, c := range counts {
		name := procedureTypeNames[c.TP]
		if name == "" {
			name = "სხვა"
		}
		categories = append(categories, ProcedureCategoryCount{
			TP:    c.TP,
			Name:  name,
			Count: c.Count,
		})
	}

	// Add allergies if any
	if allergyCount > 0 {
		categories = append(categories, ProcedureCategoryCount{
			TP:    999,
			Name:  "ალერგია / დაავადება",
			Count: int(allergyCount),
		})
	}

	writeJSON(w, http.StatusOK, PublicPetResponse{
		Pet:        info,
		Categories: categories,
	})
}